chore: ModelEditorTabContent.tsx code reformat
