+++ src/components/react-components/model-editor/ModelEditorTabContent/ModelEditorTabContent.tsx
@@ -1,29 +1,20 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useState } from 'react';
 import ModelEditor from '../../../../services/model-editor/ModelEditor/ModelEditor';
-import type { Variable } from '../../../../types';
 import DotHeaderReact from '../../lit-wrappers/DotHeaderReact';
-import SimpleHeaderReact from '../../lit-wrappers/SimpleHeaderReact';
 import TextInputReact from '../../lit-wrappers/TextInputReact';
-import VariableInfo from './VariableInfo/VariableInfo';
-import { useMemo } from 'react';
 import InvisibleInputReact from '../../lit-wrappers/InvisibleInputReact';
 import TextButtonReact from '../../lit-wrappers/TextButtonReact';
-import useVariablesStore from '../../../../stores/LiveModel/useVariablesStore';
 import ModelStatsTable from './ModelStatsTable/ModelStatsTable';
 import useModelInfoStore from '../../../../stores/LiveModel/useModelInfoStore';
 import TextIconButtonReact from '../../lit-wrappers/TextIconButtonReact';
 
 import AddIcon from '../../../../assets/icons/add_box.svg';
+import ModelEditorVariableTable from './ModelEditorVariableTable/ModelEditorVariableTable';
 
 const ModelEditorTabContent: React.FC = () => {
-  const [hoverId, setHoverId] = useState<number | null>(null);
-  const [selectedId, setSelectedId] = useState<number | null>(
-    ModelEditor.getSelectedVariableId()
-  );
   const [variableSearchText, setVariableSearchText] = useState<string>(
     ModelEditor.getVariableSearch()
   );
-
   const [showModelDescription, setShowModelDescription] =
     useState<boolean>(false);
 
@@ -31,64 +22,11 @@ const ModelEditorTabContent: React.FC = () => {
     state.getModelDescription()
   );
 
-  const variablesObj = useVariablesStore((state) => state.variables);
-  const variables = Object.values(variablesObj);
-
-  const hoverVariableInfo = useCallback((id: number, turnOnHover: boolean) => {
-    setHoverId(turnOnHover ? id : null);
-  }, []);
-
-  const selectVariableInfo = useCallback((id: number, select: boolean) => {
-    setSelectedId(select ? id : null);
-  }, []);
-
-  useEffect(() => {
-    ModelEditor.setSelectVariableFunction(selectVariableInfo);
-    ModelEditor.setHoverVariableFunction(hoverVariableInfo);
-  }, [hoverVariableInfo, selectVariableInfo]);
-
   const setVariableSearch = (name: string) => {
     if (name !== variableSearchText) {
       ModelEditor.setVariableSearch(name);
       setVariableSearchText(name);
     }
-  };
-
-  const filterVariable = (variable: Variable) => {
-    return (
-      variableSearchText === '' || variable.name.startsWith(variableSearchText)
-    );
-  };
-
-  const filteredVariables = useMemo(() => {
-    if (variableSearchText === '') return variables;
-    return variables.filter(filterVariable);
-  }, [variables, variableSearchText]);
-
-  const insertVariables = () => {
-    if (!filteredVariables || filteredVariables.length === 0) {
-      return (
-        <section className="flex h-[200px] w-[98%] justify-center items-center">
-          <SimpleHeaderReact
-            headerText="No Variables"
-            textFontWeight="normal"
-          />
-        </section>
-      );
-    }
-
-    return (
-      <section className="flex flex-col min-h-[50px] h-auto max-h-[100px] md:max-h-[200px] xl:max-h-[300px] 2xl:max-h-[400px] overflow-auto w-[98%] px-[2%] pb-1 mb-1 gap-1">
-        {filteredVariables.map((variable: Variable) => (
-          <VariableInfo
-            key={variable.id}
-            {...variable}
-            hover={hoverId === variable.id}
-            selected={selectedId === variable.id}
-          />
-        ))}
-      </section>
-    );
   };
 
   return (
@@ -157,7 +95,7 @@ const ModelEditorTabContent: React.FC = () => {
         value={variableSearchText}
       />
 
-      {insertVariables()}
+      <ModelEditorVariableTable searchText={variableSearchText} />
     </div>
   );
 };
