+++ src/components/react-components/model-editor/ModelEditorTabContent/ModelEditorVariableTable/ModelEditorVariableTableProps.ts
@@ -1,0 +1,3 @@
+export type ModelEditorVariableTableProps = {
+  searchText: string | undefined;
+};
