+++ src/components/react-components/model-editor/ModelEditorTabContent/ModelEditorVariableTable/VariableInfo/RegulationInfo/RegulationInfo.tsx
@@ -1,6 +1,6 @@
-import type { Regulation, } from '../../../../../../types';
-import ModelEditor from '../../../../../../services/model-editor/ModelEditor/ModelEditor';
-import useVariablesStore from '../../../../../../stores/LiveModel/useVariablesStore';
+import type { Regulation } from '../../../../../../../types';
+import ModelEditor from '../../../../../../../services/model-editor/ModelEditor/ModelEditor';
+import useVariablesStore from '../../../../../../../stores/LiveModel/useVariablesStore';
 
 const RegulationInfo: React.FC<Regulation> = ({
   regulator,
@@ -8,7 +8,9 @@ const RegulationInfo: React.FC<Regulation> = ({
   observable,
   monotonicity,
 }) => {
-  const regulatorVar = useVariablesStore(state => state.variableFromId(regulator));
+  const regulatorVar = useVariablesStore((state) =>
+    state.variableFromId(regulator)
+  );
 
   const getRegulationIcon = () => {
     switch (monotonicity) {
